reset password token fix

+++ src/main/java/com/reporttool/datasources/service/DataSourcePropertiesService.java
@@ -11,6 +11,8 @@ import com.reporttool.domain.exeption.ResourceNotFoundException;
 import com.reporttool.domain.model.DataSourceDbRepresentation;
 import com.reporttool.domain.model.mapper.DataSourcePropertiesMapper;
 import com.reporttool.domain.service.DataSourceDbRepresentationService;
+import com.zaxxer.hikari.HikariConfig;
+import com.zaxxer.hikari.HikariDataSource;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
@@ -18,9 +20,14 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.annotation.PostConstruct;
+import javax.sql.DataSource;
 import java.io.IOException;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.atomic.AtomicReference;
 import java.util.stream.Collectors;
 
 import static java.util.Objects.nonNull;
@@ -35,12 +42,31 @@ public class DataSourcePropertiesService {
     private final ObjectMapper objectMapper;
     private final CipherService cipherService;
 
+    /*
+     * Field to store initialized DataSources
+     */
+    private final Map<String, AtomicReference<DataSource>> dataSources = new ConcurrentHashMap<>();
+
+    /**
+     * Read from data base all stored {@link DataSource} properties
+     * and initialize them
+     */
+    @PostConstruct
+    private void init() {
+        List<DataSourceProperties> dataSourceProperties = findAll();
+        dataSourceProperties.forEach(a -> dataSources.put(a.getDataSourceName(), createDataSourceReference(a)));
+    }
 
 
+
+    @Transactional
     public DataSourceForm saveDataSource(DataSourceForm dataSourceForm) {
         DataSourceProperties dbProperties = dataSourcePropertiesMapper.mapToDataSourceProperties(dataSourceForm);
         DataSourceDbRepresentation createdDbRepresentation =
                 dbRepresentationService.create(createDbRepresentation(dbProperties));
+
+        addDataSource(dbProperties);
+
         String decodedStringDbPropertiesRepresentation =
                 cipherService.decrypt(createdDbRepresentation.getDataSourceObjectRepresentation());
         return dataSourcePropertiesMapper
@@ -50,7 +76,7 @@ public class DataSourcePropertiesService {
 
 
 
-    public List<DataSourceProperties> findAll() {
+    private List<DataSourceProperties> findAll() {
         return mapToDataSourceForm(dbRepresentationService.findAll());
     }
 
@@ -96,14 +122,22 @@ public class DataSourcePropertiesService {
         dbRepresentation.setDataSourceObjectRepresentation(encryptedUpdatedStringDataBasePropertiesRepresentation);
 
         dbRepresentationService.update(dbRepresentation);
+
+        updateDataSource(dbProperties);
+
         log.debug("DataSourceDbRepresentation was successfully updated");
         return dataSourcePropertiesMapper.mapToDataSourceEditForm(dbProperties);
     }
 
 
 
+    @Transactional
     public void deleteDataSource(Long dataSourceId) {
-        dbRepresentationService.delete(dataSourceId);
+        Optional<DataSourceDbRepresentation> optionalDbRepresentation = dbRepresentationService.findById(dataSourceId);
+        if (optionalDbRepresentation.isPresent()) {
+            dbRepresentationService.delete(dataSourceId);
+            deleteDataSource(optionalDbRepresentation.get().getDataSourceName());
+        }
     }
 
 
@@ -173,5 +207,34 @@ public class DataSourcePropertiesService {
         if (nonNull(dataSourceForm.getDataSourceType())) {
             dbProperties.setDataSourceType(dataSourceForm.getDataSourceType());
         }
+    }
+
+    private AtomicReference<DataSource> createDataSourceReference(DataSourceProperties dbProperties) {
+        AtomicReference<DataSource> atomicReference = new AtomicReference<>();
+        atomicReference.set(createDataSource(dbProperties));
+        return atomicReference;
+    }
+
+    private DataSource createDataSource(DataSourceProperties dataSourceProperties) {
+        HikariConfig config = new HikariConfig();
+        config.setUsername(dataSourceProperties.getUserName());
+        config.setPassword(dataSourceProperties.getPassword());
+        config.setJdbcUrl(dataSourceProperties.getUrl());
+        config.setDriverClassName(dataSourceProperties.getDriverClassName());
+        return new HikariDataSource(config);
+    }
+
+    private void addDataSource(DataSourceProperties dbProperties) {
+        AtomicReference<DataSource> atomicReference = new AtomicReference<>();
+        atomicReference.set(createDataSource(dbProperties));
+        dataSources.put(dbProperties.getDataSourceName(), atomicReference);
+    }
+
+    private void updateDataSource(DataSourceProperties dbProperties) {
+        dataSources.get(dbProperties.getDataSourceName()).set(createDataSource(dbProperties));
+    }
+
+    private void deleteDataSource(String dataSourceName) {
+        dataSources.remove(dataSourceName);
     }
 }
